Display messages that consist of only a link to an image as an image (when the image preview is available and visible).
